% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSelect.r
\name{check_object}
\alias{check_object}
\title{Check Object class}
\usage{
check_object(x)
}
\arguments{
\item{x}{A timeseries object}
}
\value{
stops if object not a ts class
}
\description{
Internal function that verifies the class of the object
 (should be time series)
}
\details{
internal function for verifying that the object belongs
 to class "time series"
}
\author{
Avi Blinder
}

