\name{ros2_ts}
\alias{ros2_ts}
\docType{data}
\title{"Time Series sample 2" }
\description{This sample dataset is taken from a Kaggle's competition ("Rossman Store Sales")}
\usage{data("ros2_ts")}
\format{
  The format is:
 Time-Series [1:365] from 1 to 365: 0.1 5530 4327 4486 4997 ...
}
\details{Sales data from the second store are represented in the dataset}
\source{"https://www.kaggle.com/c/rossmann-store-sales/data"}
\examples{
data(ros2_ts)
}
\keyword{datasets}
