% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSelect.r
\name{run_models}
\alias{run_models}
\title{Run ts Models}
\usage{
run_models(ts1, accuracy_measure = NULL)
}
\arguments{
\item{ts1}{A timeseries object}

\item{accuracy_measure}{- Possilbe error meassures:
ME, RMSE, MAE, MPE ,MAPE, MASE, ACF1}
}
\value{
the best time series model
}
\description{
Function that executes several models and picks
the best one.
}
\examples{
data(ros1_ts)
run_models(ros1_ts)
run_models(ros1_ts,"RMSE")
}
\author{
Avi Blinder
}

