\name{ros1_ts}
\alias{ros1_ts}
\docType{data}
\title{"Time Series sample" }
\description{This sample dataset is taken from a Kaggle's competition ("Rossman Store Sales")}
\usage{data("ros1_ts")}
\format{
  The format is:
 Time-Series [1:942] from 1 to 942: 0.1 5530 4327 4486 4997 ...
}
\details{Only the sales data from the first store is represented in the dataset}
\source{"https://www.kaggle.com/c/rossmann-store-sales/data"}
\examples{
data(ros1_ts)
## maybe str(ros1_ts) ; plot(ros1_ts) ...
}
\keyword{datasets}
